requirements.txt
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.25.0
pandas>=2.0.0
plotly>=5.18.0
python-docx>=1.1.0
PyPDF2>=3.0.0
lxml>=5.0.0
reportlab>=4.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
